refactor: split api.ts into types, the window bridge, and formatting helpers
